Adapt tests to new __utils__ form
