+++ srv/modules/utils/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+# vim: ts=8 et sw=4 sts=4
